Auto heartbeat for sync interface (#99)
Auto heartbeat for C/C++ is now complete. Only Qt remains.

Also, fixed some potential race conditions and cleaned up some
internal interfaces.

+++ windows/pb_sleep_ms.c
@@ -1,7 +1,7 @@
 /* -*- c-file-style:"stroustrup"; indent-tabs-mode: nil -*- */
 #include <windows.h>
 
-void pb_sleep_ms(DWORD dwMilliseconds)
+void pb_sleep_ms(unsigned long dwMilliseconds)
 {
     Sleep(dwMilliseconds);
 }
